+++ core/src/main/scala/org/broadinstitute/dsp/workbench/welder/package.scala
@@ -7,17 +7,18 @@ import java.util.Base64
 
 import cats.Eq
 import cats.effect.concurrent.Ref
-import cats.effect.{Concurrent, ContextShift, IO, Sync}
+import cats.effect.{Concurrent, ContextShift, IO, Resource, Sync}
 import cats.implicits._
 import fs2.{Pipe, Stream}
 import io.chrisdavenport.log4cats.Logger
-import io.circe.Decoder
+import io.circe.{Decoder, Encoder, Printer}
 import io.circe.fs2._
 import org.broadinstitute.dsde.workbench.google2.GcsBlobName
 import org.broadinstitute.dsde.workbench.model.WorkbenchEmail
 import org.broadinstitute.dsde.workbench.model.google.GcsBucketName
 import org.broadinstitute.dsp.workbench.welder.SourceUri.GsPath
 import org.typelevel.jawn.AsyncParser
+import io.circe.syntax._
 import scala.language.implicitConversions
 import scala.concurrent.ExecutionContext
 import scala.concurrent.ExecutionContext.global
@@ -91,6 +92,26 @@ package object welder {
   type MetadataCache = Ref[IO, Map[RelativePath, AdaptedGcsMetadataCache]]
 
   val gcpObjectType = "text/plain"
+
+  def cachedResource[A, B: Decoder: Encoder](path: Path, blockingEc: ExecutionContext, toTuple: B => List[(A, B)])(
+    implicit logger: Logger[IO], cs: ContextShift[IO]
+  ): Stream[IO, Ref[IO, Map[A, B]]] =
+    for {
+      cached <- readJsonFileToA[IO, List[B]](path).map(ls => ls.flatMap(b => toTuple(b)).toMap).handleErrorWith { error =>
+        Stream.eval(logger.info(s"$path not found")) >> Stream.emit(Map.empty[A, B]).covary[IO]
+      }
+      ref <- Stream.resource(
+        Resource.make(Ref.of[IO, Map[A, B]](cached))(
+          ref =>
+            Stream
+              .eval(ref.get)
+              .flatMap(x => Stream.emits(x.values.toSet.asJson.pretty(Printer.noSpaces).getBytes("UTF-8")))
+              .through(fs2.io.file.writeAll(path, blockingEc))
+              .compile
+              .drain
+        )
+      )
+    } yield ref
 
   implicit val eqLocalDirectory: Eq[LocalDirectory] = Eq.instance((p1, p2) => p1.path.toString == p2.path.toString)
   implicit def loggerToContextLogger[F[_]](logger: Logger[F]): ContextLogger[F] = ContextLogger(logger)
